requirements.txt
channels==4.0.0
Django==4.2.7
djangorestframework==3.14.0
dnspython==2.4.2
geoip2==4.7.0
phonenumbers==8.13.26
pycountry==22.3.5
python-dotenv==1.0.0
Requests==2.31.0
sentry_sdk==1.38.0
uvicorn
uvloop
django-celery-beat
celery
django-cors-headers
whitenoise
gevent
djangorestframework-camel-case
beautifulsoup4
python-dotenv

# API Key
djangorestframework-api-key

# Cache
django-redis

## GeoIP
geoip2
pycountry
phonenumbers
django-ipware